Port to NaCl

+++ DROD/GameScreen.cpp
@@ -4109,7 +4109,7 @@ SCREENTYPE CGameScreen::ProcessCueEventsAfterRoomDraw(
 	if (CueEvents.HasOccurred(CID_RoomConquerPending))  //priority of temporary moods
 	{
 		UINT eClearID = GetPlayerClearSEID();
-		if (eClearID == SEID_NONE) eClearID = SEID_CLEAR;
+		if (eClearID == (UINT)SEID_NONE) eClearID = SEID_CLEAR;
 		this->pFaceWidget->SetMoodToSoundEffect(Mood_Happy, SEID(eClearID));
 	}
 	else if (CueEvents.HasOccurred(CID_MonsterDiedFromStab))
